Подстроил под сборку
Подстроил всю работу под свою сборку Svelte.
