+++ .editorconfig
@@ -1,0 +1,7 @@
+# editorconfig.org
+root = true
+
+[*]
+charset = utf-8
+insert_final_newline = true
+trim_trailing_whitespace = true
