+++ .prettierrc
@@ -1,0 +1,8 @@
+{
+  "endOfLine": "lf",
+  "printWidth": 88,
+  "semi": true,
+  "singleQuote": false,
+  "tabWidth": 2,
+  "trailingComma": "es5"
+}
