+++ README.md
@@ -1,1 +1,1 @@
-
+# Замечательная сборочка для работы со Svelte!
