--- index.kit
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-	<meta charset="utf-8">
-	<meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=0">
-	<meta name="format-detection" content="telephone=no">
-	<title>Backstory popup</title>
-</head>
-<body class="body">
-    <div class="wrapper">
-        <!-- @import "backstory-popup.html" -->
-    </div>
-</body>
-</html>
