+++ package.json
@@ -1,0 +1,32 @@
+{
+  "name": "svelte-app",
+  "version": "1.4.9",
+  "scripts": {
+    "build": "rollup -c",
+    "dev": "rollup -c -w",
+    "start": "sirv public",
+    "format": "npx prettier --write \"src/{,!(node_modules)/**/}*.{js,svelte,html}\""
+  },
+  "devDependencies": {
+    "@rollup/plugin-commonjs": "^22.0.2",
+    "@rollup/plugin-json": "^4.1.0",
+    "@rollup/plugin-node-resolve": "^14.1.0",
+    "autoprefixer": "^10.4.19",
+    "eslint": "^8.57.0",
+    "eslint-plugin-svelte3": "^4.0.0",
+    "postcss": "^8.4.38",
+    "prettier": "^2.8.8",
+    "prettier-plugin-svelte": "^2.10.1",
+    "rollup": "^2.79.1",
+    "rollup-plugin-livereload": "^2.0.5",
+    "rollup-plugin-postcss": "^4.0.2",
+    "rollup-plugin-svelte": "^7.2.0",
+    "rollup-plugin-terser": "^7.0.2",
+    "svelte": "^3.59.2",
+    "svelte-routing": "^2.12.0",
+    "tailwindcss": "^2.2.19"
+  },
+  "dependencies": {
+    "sirv-cli": "^2.0.2"
+  }
+}
