+++ prepros-6.config
@@ -1,20 +1,21 @@
 {
-  "name": "TaxiMoney-BackstoryPopup",
+  "name": "svelte-app",
   "firstRun": false,
   "exportConfig": true,
   "fileConfigs": [
     {
-      "path": "styles/backstory-popup.scss",
-      "configJson": "{\"forceCompile\":false,\"customOutput\":\"styles/backstory-popup.css\",\"autoCompile\":true,\"sourceMap\":false,\"compiler-node-sass\":{\"enabled\":true,\"outputStyle\":\"nested\"},\"compiler-autoprefixer\":{\"enabled\":true},\"compiler-minify-css\":{\"enabled\":true},\"outputSuffix\":null}"
+      "path": "src/scss/main.scss",
+      "configJson": "{\"forceCompile\":false,\"customOutput\":\"public/build/main.min.css\",\"autoCompile\":true,\"sourceMap\":false,\"compiler-node-sass\":{\"enabled\":true,\"outputStyle\":\"nested\"},\"compiler-autoprefixer\":{\"enabled\":true},\"compiler-minify-css\":{\"enabled\":true},\"outputSuffix\":null}"
     },
     {
-      "path": "index.kit",
-      "configJson": "{\"forceCompile\":false,\"customOutput\":\"index.html\",\"autoCompile\":true,\"compiler-kit\":{\"enabled\":true},\"outputSuffix\":null}"
+      "path": "src/scss/_backstory-popup.scss",
+      "configJson": "{\"forceCompile\":false,\"customOutput\":\"public/build/main.min.css\",\"autoCompile\":true,\"sourceMap\":false,\"compiler-node-sass\":{\"enabled\":true,\"outputStyle\":\"nested\"},\"compiler-autoprefixer\":{\"enabled\":true},\"compiler-minify-css\":{\"enabled\":true},\"outputSuffix\":null}"
     }
   ],
   "fileTree": {
     "expandedDirs": [
-      "scripts"
+      "src",
+      "src/scss"
     ],
     "hideSystemFiles": true,
     "systemFiles": [
@@ -28,14 +29,7 @@
     ],
     "hideUnwatchedFiles": false
   },
-  "imports": [
-    {
-      "path": "index.kit",
-      "imports": [
-        "backstory-popup.html"
-      ]
-    }
-  ],
+  "imports": [],
   "projectView": {
     "selectedView": "file-tree"
   },
@@ -78,7 +72,7 @@
     "wp-includes"
   ],
   "server": {
-    "port": 7879,
+    "port": 5174,
     "assignNewPortAutomatically": true,
     "enable": true,
     "proxy": {
