+++ src/BackstoryPopup.svelte
@@ -1,0 +1,326 @@
+<!-- * Скрипт для окна предыстории -->
+<script>
+    import BackstoryPopupScript from './scripts/BackstoryPopup.svelte';
+</script>
+<BackstoryPopupScript />
+
+<!-- * Само окно предыстории -->
+<!-- Оболочка окна -->
+<div class="backstory-popup-wrap">
+    <!-- Окно предыстории -->
+    <div class="backstory-popup">
+        <!--
+        — — — — — — — — — — — — — — — — — — — — — —
+        *                                       Заставка
+        — — — — — — — — — — — — — — — — — — — — — —
+        -->
+        <div class="backstory-popup__title-screen">
+            <!-- Логотип -->
+            <img src="assets/taxi-money-logo.svg" alt="" class="backstory-popup__title-screen-logo">
+        </div>
+        <!--
+        — — — — — — — — — — — — — — — — — — — — — —
+        *                               Содержимое окна
+        — — — — — — — — — — — — — — — — — — — — — —
+        -->
+        <div class="backstory-popup__content">
+            <!--
+            — — — — — — — — — — — — — — — — — — — — — —
+            *                               1) Такси подъехало
+            — — — — — — — — — — — — — — — — — — — — — —
+            -->
+            <div class="backstory-popup__slide backstory-popup__slide-1">
+                <!-- Фоновое видео -->
+                <video class="backstory-popup__video" width="100%" autoplay loop muted playsinline>
+                    <source src="assets/01.webm" type="video/webm">
+                </video>
+                <!-- Пузырь реплики -->
+                <!-- ! Отсутствует -->
+                <!--Рассказчик (разбит на 2 линии) -->
+                <div class="backstory-popup__narrator">
+                    <!-- Слова рассказчика -->
+                    <div class="backstory-popup__narrator-row-1">
+                        <span class="backstory-popup__typing-string-wrap">
+                            <span class="backstory-popup__typing-string--1s">
+                            Быстро машина подъехала! Супер!
+                            </span>
+                        </span>
+                        <span class="backstory-popup__typing-string-wrap">
+                            <span class="backstory-popup__typing-string--2s">
+                                Похоже, сегодня какой-то особенный день! Всё получается так легко!
+                            </span>
+                        </span>
+                    </div>
+                    <!-- Кнопки и формы -->
+                    <div class="backstory-popup__narrator-row-2">
+                        <button class="backstory-popup__btn--blue">Сесть в такси</button>
+                    </div>
+                </div>
+            </div>
+
+
+            <!--
+            — — — — — — — — — — — — — — — — — — — — — —
+            *                          2) Нападение грабителей
+            — — — — — — — — — — — — — — — — — — — — — —
+            -->
+            <div class="backstory-popup__slide backstory-popup__slide-2">
+                <!-- Фоновое видео -->
+                <video class="backstory-popup__video" width="100%" autoplay loop muted playsinline>
+                    <source src="assets/02.webm" type="video/webm">
+                </video>
+                <!-- Элемент для эффекта фотографирования -->
+                <div class="backstory-popup__slide-photographing-screen"></div>
+                <!-- Элемент для анимации вызова полиции-->
+                <div class="backstory-popup__slide-police-calling-screen"></div>
+                <!-- Пузырь реплики -->
+                <!-- ! Отсутствует -->
+                <!--Рассказчик (разбит на 2 линии) -->
+                <div class="backstory-popup__narrator">
+                    <!-- Слова рассказчика -->
+                    <div class="backstory-popup__narrator-row-1">
+                        <span class="backstory-popup__typing-string-wrap">
+                            <span class="backstory-popup__typing-string--2s">
+                                Такси не успело проехать и пары сотен метров, как дорогу заблокировала
+                            </span>
+                        </span>
+                        <span class="backstory-popup__typing-string-wrap">
+                            <span class="backstory-popup__typing-string--2s">
+                                другая машина. Таксиста грабили вооруженные люди в масках!
+                            </span>
+                        </span>
+                    </div>
+                    <!-- Кнопки и формы -->
+                    <div class="backstory-popup__narrator-row-2">
+                        <button class="backstory-popup__btn--orange">Заступиться</button>
+                        <button class="backstory-popup__btn--orange">Сфотографировать преступников</button>
+                        <button class="backstory-popup__btn--orange">Вызвать полицию</button>
+                    </div>
+                </div>
+            </div>
+
+
+            <!--
+            — — — — — — — — — — — — — — — — — — — — — —
+            *                          3) Угроза  грабителей
+            — — — — — — — — — — — — — — — — — — — — — —
+            -->
+            <div class="backstory-popup__slide backstory-popup__slide-3">
+                <!-- Фоновое видео -->
+                <video class="backstory-popup__video" width="100%" autoplay loop muted playsinline>
+                    <source src="assets/03.webm" type="video/webm">
+                </video>
+                <!-- Элемент для анимации видеозаписи -->
+                <div class="backstory-popup__slide-recording-screen"></div>
+                <!-- Элемент для эффекта фотосъёмки-->
+                <div class="backstory-popup__slide-photographing-screen"></div>
+                <!-- Пузырь реплики -->
+                <div class="backstory-popup__bubble">
+                    <span class="backstory-popup__typing-string-wrap">
+                        <span class="backstory-popup__typing-string--1s">
+                            Копы! Валим!!! Эй, ты! Держи язык за
+                        </span>
+                    </span>
+                    <span class="backstory-popup__typing-string-wrap">
+                        <span class="backstory-popup__typing-string--1s">
+                            зубами, ясно! Иначе примеришь
+                        </span>
+                    </span>
+                    <span class="backstory-popup__typing-string-wrap">
+                        <span class="backstory-popup__typing-string--1s">
+                            колумбийский галстук! 🔪 💀 🔫
+                        </span>
+                    </span>
+                    <span class="backstory-popup__buble-triangle"></span>
+                </div>
+
+                <!--Рассказчик (разбит на 2 линии) -->
+                <div class="backstory-popup__narrator">
+                    <!-- Слова рассказчика -->
+                    <!-- ! Отсутствует -->
+                    <!-- Кнопки и формы -->
+                    <div class="backstory-popup__narrator-row-2">
+                        <button class="backstory-popup__btn--orange">Снять преступников на видео</button>
+                        <button class="backstory-popup__btn--orange">Сфотографировать их машину</button>
+                        <button class="backstory-popup__btn--orange">Попытаться их задержать</button>
+                    </div>
+                </div>
+            </div>
+
+
+            <!--
+            — — — — — — — — — — — — — — — — — — — — — —
+            *                          4) В полицейском участке
+            — — — — — — — — — — — — — — — — — — — — — —
+            -->
+            <div class="backstory-popup__slide backstory-popup__slide-4">
+                <!-- Фоновое видео -->
+                <video class="backstory-popup__video" width="100%" autoplay loop muted playsinline>
+                    <source src="assets/04.webm" type="video/webm">
+                </video>
+                <!-- Пузыри реплик -->
+                <div class="backstory-popup__bubble">
+                    <span class="backstory-popup__typing-string-wrap">
+                        <span class="backstory-popup__typing-string--1s">
+                            Благодарю Вас за помощь! Надеюсь, предоставленные
+                        </span>
+                    </span>
+                    <span class="backstory-popup__typing-string-wrap">
+                        <span class="backstory-popup__typing-string--1s">
+                            Вами материалы помогут на поймать этих отморозков.
+                        </span>
+                    </span>
+                    <span class="backstory-popup__typing-string-wrap">
+                        <span class="backstory-popup__typing-string--1s">
+                            Это уже не первое их преступление. И они очень опасны.
+                        </span>
+                    </span>
+                    <span class="backstory-popup__typing-string-wrap">
+                        <span class="backstory-popup__typing-string--1s">
+                                Мы настаиваем на программе защиты свидетелей!
+                        </span>
+                    </span>
+
+                    <span class="backstory-popup__buble-triangle"></span>
+                </div>
+
+                <div class="backstory-popup__bubble">
+                    <span class="backstory-popup__typing-string-wrap">
+                        <span class="backstory-popup__typing-string--1s">
+                            Мы отправим Вас в Такси-Полис. Поработаете таксистом,
+                        </span>
+                    </span>
+                    <span class="backstory-popup__typing-string-wrap">
+                        <span class="backstory-popup__typing-string--1s">
+                            пока всё не уляжется. Сможете зарабатывать <b class="backstory-popup__bold--pink">до 6000 ₽</b> в
+                        </span>
+                    </span>
+                    <span class="backstory-popup__typing-string-wrap">
+                        <span class="backstory-popup__typing-string--1s">
+                            день! Друзей новых найдёте. Там хорошо! 👍
+                        </span>
+                    </span>
+
+                    <span class="backstory-popup__buble-triangle"></span>
+                </div>
+
+                <!--Рассказчик (разбит на 2 линии) -->
+                <div class="backstory-popup__narrator">
+                    <!-- Слова рассказчика -->
+                    <div class="backstory-popup__narrator-row-1">
+                        <span class="backstory-popup__typing-string-wrap">
+                            <span class="backstory-popup__typing-string--2s">
+                                Придумай себе кодовое имя для программы защиты свидетелей.
+                            </span>
+                        </span>
+                    </div>
+                    <!-- Кнопки и формы -->
+                    <div class="backstory-popup__narrator-row-2">
+                        <form action="" class="backstory-popup__form">
+                            <input type="text" class="backstory-popup__input" placeholder="Введи новое имя">
+                            <input type="submit" class="backstory-popup__btn--blue" value="Сохранить имя">
+                        </form>
+                    </div>
+                </div>
+            </div>
+
+
+            <!--
+            — — — — — — — — — — — — — — — — — — — — — —
+            *                          5) Получение паспорта
+            — — — — — — — — — — — — — — — — — — — — — —
+            -->
+            <div class="backstory-popup__slide backstory-popup__slide-5">
+                <!-- Фоновое видео -->
+                <video class="backstory-popup__video" width="100%" autoplay loop muted playsinline>
+                    <source src="assets/05.webm" type="video/webm">
+                </video>
+                <!-- Пузырь реплики -->
+                <!-- ! Отсутствует -->
+                <!--Рассказчик (разбит на 2 линии) -->
+                <div class="backstory-popup__narrator">
+                    <!-- Слова рассказчика -->
+                    <div class="backstory-popup__narrator-row-1">
+                        <span class="backstory-popup__typing-string-wrap">
+                            <span class="backstory-popup__typing-string--2s">
+                                Паспорт жителя Taxi-Polis будет доставлен тебе по электронной
+                            </span>
+                        </span>
+                        <span class="backstory-popup__typing-string-wrap">
+                            <span class="backstory-popup__typing-string--1s">
+                                почте. Укажи свой e-mail.
+                            </span>
+                        </span>
+                    </div>
+                    <!-- Кнопки и формы -->
+                    <div class="backstory-popup__narrator-row-2">
+                        <form action="" class="backstory-popup__form">
+                            <input type="text" class="backstory-popup__input" placeholder="Введи свой e-mail">
+                            <input type="submit" class="backstory-popup__btn--orange" value="Получить паспорт">
+                        </form>
+                    </div>
+                </div>
+            </div>
+
+
+            <!--
+            — — — — — — — — — — — — — — — — — — — — — —
+            *                          6) Встреча с Марго
+            — — — — — — — — — — — — — — — — — — — — — —
+            -->
+            <div class="backstory-popup__slide backstory-popup__slide-6">
+                <!-- Кнопка перезапуска слайдера окна -->
+                <button class="backstory-popup__btn-restart">Пройти снова</button>
+                <!-- Фоновое видео -->
+                <video class="backstory-popup__video" width="100%" autoplay loop muted playsinline>
+                    <source src="assets/06.webm" type="video/webm">
+                </video>
+                <!-- Пузырь реплики -->
+                <div class="backstory-popup__bubble">
+                    <span class="backstory-popup__typing-string-wrap">
+                        <span class="backstory-popup__typing-string--1s">
+                            Привет, <span class="backstory-popup__bold--pink backstory-popup__username">Samiznaetekto78</span>! Меня зовут Марго, и я
+                        </span>
+                    </span>
+                    <span class="backstory-popup__typing-string-wrap">
+                        <span class="backstory-popup__typing-string--1s">
+                            буду помогать тебе устроиться в Такси-Полисе.
+                        </span>
+                    </span>
+                    <span class="backstory-popup__typing-string-wrap">
+                        <span class="backstory-popup__typing-string--1s">
+                            Скучно не будет! Начнём? 😉
+                        </span>
+                    </span>
+
+                    <span class="backstory-popup__buble-triangle"></span>
+                </div>
+                <!--Рассказчик (разбит на 2 линии) -->
+                <div class="backstory-popup__narrator">
+                    <!-- Слова рассказчика -->
+                    <div class="backstory-popup__narrator-row-1">
+                        <span class="backstory-popup__typing-string-wrap">
+                            <span class="backstory-popup__typing-string--2s">
+                                Поздравляем! Теперь у тебя есть новый паспорт, и ты можешь начать
+                            </span>
+                        </span>
+                        <span class="backstory-popup__typing-string-wrap">
+                            <span class="backstory-popup__typing-string--2s">
+                                жизнь с чистого листа в Taxi-Polis. Впереди много захватывающих
+                            </span>
+                        </span>
+                        <span class="backstory-popup__typing-string-wrap">
+                            <span class="backstory-popup__typing-string--1s">
+                                приключений и классных возможностей!
+                            </span>
+                        </span>
+                    </div>
+                    <!-- Кнопки и формы -->
+                    <div class="backstory-popup__narrator-row-2">
+                        <button class="backstory-popup__btn--orange">Покорить Такси-Полис! 🚀</button>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </div>
+</div>
