+++ src/backstory-popup.js
@@ -1,0 +1,7 @@
+import BackstoryPopup from "./BackstoryPopup.svelte";
+
+const backstoryPopup = new BackstoryPopup({
+  target: document.body
+});
+
+export default backstoryPopup;
