+++ src/static/index.html
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html>
+<head>
+	<meta charset="utf-8">
+	<meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=0">
+	<meta name="format-detection" content="telephone=no">
+	<title>Backstory popup</title>
+  <script defer src="build/bundle.js"></script>
+  <link rel="stylesheet" href="build/main.min.css">
+</head>
+
+  <body></body>
+</html>
