+++ src/static/manifest.json
@@ -1,0 +1,16 @@
+{
+  "name": "svelte-app",
+  "short_name": "svelte-app",
+  "scope": "/%base%",
+  "start_url": "/%base%index.html",
+  "background_color": "#ffffff",
+  "theme_color": "#000000",
+  "display": "minimal-ui",
+  "icons": [
+    {
+      "src": "logo.png",
+      "sizes": "512x512",
+      "type": "image/png"
+    }
+  ]
+}
